Fix #11 put JSONMessage in separat file
